refactor: better variable name & reorganize object wrapping
I change the way of property accessing since it looks not common.

For instance, "person1.firstname" is often used to imply "the firstname
of person1". So I reorganize the object in order to access properties in
this way.

Also, the variable name "item" is to vague. It is used to store
different types of message for different players. So I give those
properties better names as well.