refactor: no 'new' for side effects
The `new` keyword impies the creation of a new object. If the `Game`
class is declared in the ES5 function way, I could just remove the `new`
keyword to impies that this statement might have side effects. However,
I couldn't. So I declare a new variable for storing this `Game` instance
and call the `game.start();` manually.

Also, with this way, one can decided when to start the game more freely.

(eslint: no-new)